refactor(batch): decompose replay.rs into submodules (SPL-23 #2451)
Split the 1330 LoC monolithic replay.rs into focused submodules under
crates/batch/src/replay/ to satisfy POST-8 audit LoC budget (<650 per
file). Public API (replay, apply_delta_ops, ReplayResult) is unchanged.

- mod.rs: orchestrator, ReplayResult, phase 1/3 helpers (260 LoC)
- codec.rs: CompressionCodec, decoder factory, zstd auto-detect (202)
- delta.rs: apply_delta_ops, write_literals_to_file, choose_block_length (216)
- delta_phase.rs: NDX-stream loop, codec state machine, per-file dispatch (467)
- dispatch.rs: iflags/sum_head/checksum readers, streaming-zlib loop (243)
- fs_ops.rs: cross-platform symlink + metadata application (96)
- tests/: unit tests for delta primitives and decoder construction (204)

Behavioural diff vs original: none. The compressed-codec auto-detect
state machine is now encapsulated in CodecState; the swap from zlib to
zstd happens once in detect_once() before the first token read, matching
the original control flow exactly.
