+++ crates/batch/src/replay/fs_ops.rs
@@ -1,0 +1,96 @@
+//! Filesystem helpers for batch replay: symlink creation and metadata application.
+//!
+//! Symlink creation has platform-specific handling: Unix uses
+//! `std::os::unix::fs::symlink`, Windows uses `symlink_file`, and any other
+//! target returns [`BatchError::Unsupported`]. Metadata application delegates
+//! to the `metadata` crate so that permissions, timestamps, and ownership are
+//! applied with the same semantics as a live transfer.
+
+use std::fs;
+use std::path::Path;
+
+use crate::error::{BatchError, BatchResult};
+use crate::format::BatchFlags;
+
+/// Apply metadata (permissions, timestamps) from a protocol file entry to a
+/// destination path.
+///
+/// Uses the `metadata` crate's [`metadata::apply_metadata_from_file_entry`]
+/// to set permissions and modification times on the target file or directory.
+/// Ownership is applied only when the corresponding batch flags are set.
+///
+/// # Errors
+///
+/// Returns [`BatchError::Io`] if metadata cannot be applied.
+pub(super) fn apply_entry_metadata(
+    dest_path: &Path,
+    entry: &protocol::flist::FileEntry,
+    flags: &BatchFlags,
+) -> BatchResult<()> {
+    let options = metadata::MetadataOptions::new()
+        .preserve_permissions(true)
+        .preserve_times(true)
+        .preserve_owner(flags.preserve_uid)
+        .preserve_group(flags.preserve_gid);
+
+    metadata::apply_metadata_from_file_entry(dest_path, entry, &options).map_err(|e| {
+        BatchError::Io(std::io::Error::other(format!(
+            "failed to apply metadata to '{}': {e}",
+            dest_path.display()
+        )))
+    })?;
+
+    Ok(())
+}
+
+/// Create a symlink at `dest_path` pointing to the given `target`.
+///
+/// On Unix, creates a symbolic link. On other platforms, falls back to
+/// file copy (symlink creation is platform-specific).
+#[cfg(unix)]
+pub(super) fn create_symlink(target: &Path, dest_path: &Path) -> BatchResult<()> {
+    // Remove existing entry if present, to mirror upstream rsync behavior
+    if dest_path.exists() || dest_path.symlink_metadata().is_ok() {
+        let _ = fs::remove_file(dest_path);
+    }
+    std::os::unix::fs::symlink(target, dest_path).map_err(|e| {
+        BatchError::Io(std::io::Error::new(
+            e.kind(),
+            format!(
+                "failed to create symlink '{}' -> '{}': {e}",
+                dest_path.display(),
+                target.display()
+            ),
+        ))
+    })
+}
+
+/// Create a symlink on Windows (best-effort directory detection).
+#[cfg(not(unix))]
+pub(super) fn create_symlink(target: &Path, dest_path: &Path) -> BatchResult<()> {
+    if dest_path.exists() || dest_path.symlink_metadata().is_ok() {
+        let _ = fs::remove_file(dest_path);
+    }
+    // Windows requires knowing whether the target is a file or directory.
+    // Default to file symlink; directory symlinks are rare in rsync batch use.
+    #[cfg(windows)]
+    {
+        std::os::windows::fs::symlink_file(target, dest_path).map_err(|e| {
+            BatchError::Io(std::io::Error::new(
+                e.kind(),
+                format!(
+                    "failed to create symlink '{}' -> '{}': {e}",
+                    dest_path.display(),
+                    target.display()
+                ),
+            ))
+        })
+    }
+    #[cfg(not(windows))]
+    {
+        let _ = (target, dest_path);
+        Err(BatchError::Unsupported(
+            "symlink creation not supported on this platform".to_owned(),
+        ))
+    }
+}
